desafio completo

+++ src/services/product.service.js
@@ -53,8 +53,24 @@ class PdctService {
         }
     }
 
-    async createOne(title, description, code, price, stock, status, thumbnails) {
+    async createOne(title, description, code, price, stock, status, thumbnails, dataUser) {
         try {
+            const { email, role } = dataUser;
+            console.log(role);
+            let owner;
+
+            // Verificar el rol del propietario
+            if (role === 'premium') {
+                // Si el propietario es premium, establece el propietario como su email
+                owner = email;
+            } else if (role === 'admin') {
+                // Si el propietario es admin, establece el propietario como "admin"
+                owner = 'admin';
+            } else {
+                // Si el rol del propietario no es ni premium ni admin, devuelve un código de estado 403 (Prohibido)
+                throw new Error("No tienes permisos para crear productos.");
+            }
+
             this.validateProduct(description, title, code, price, stock, status);
 
             const productCreated = await ProductManager.createOne({
@@ -65,6 +81,7 @@ class PdctService {
                 stock,
                 status,
                 thumbnails,
+                owner,
             });
 
             return productCreated;
@@ -74,9 +91,36 @@ class PdctService {
         }
     }
 
-    async deleteOne(id) {
-        const deleted = await ProductManager.deleteOne({ _id: id });
-        return deleted;
+    async deleteOne(id, data) {
+        const { role, ownerId } = data;
+    
+        try {
+            // Obtener el producto que se va a eliminar utilizando la función getOne
+            const productToDelete = await this.getOne(id);
+            
+            if (role === 'admin') {
+                // Si el usuario es un administrador, puede borrar cualquier producto
+                const deleted = await ProductManager.deleteOne({ _id: id });
+                return deleted;
+            } else if (role === 'premium') {
+                // Si el usuario es premium, verifica si el producto le pertenece antes de borrarlo
+                if (productToDelete.owner === ownerId) {
+                    console.log("se cumple");
+                    const deleted = await ProductManager.deleteOne({ _id: id });
+                    return deleted;
+                } else {
+                    // Si el producto no le pertenece, devuelve un código de estado 403 (Prohibido)
+                    throw new Error("No tienes permisos para borrar este producto.");
+                }
+            } else {
+                // Si el rol del usuario no es ni admin ni premium, devuelve un código de estado 403 (Prohibido)
+                throw new Error("No tienes permisos para borrar productos.");
+            }
+        } catch (error) {
+            // Manejar errores al obtener el producto
+            console.error(error);
+            throw new Error("Ha ocurrido un error al buscar o borrar el producto");
+        }
     }
 
     async updateOne(id, description, title, code, price, stock, status, thumbnails) {
